Automated Text Extraction/Streamlit/requirements.txt
streamlit
requests
pandas
boto3
python-dotenv